Add partial support for shared crosshair
Panel can now revieve shared crosshair events and highlight images.

Also refector to use function instead of classes.
