+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.6.0
+
+- Partial support (it can only receive events) for "shared crosshair". Image background will change for the configured
+colored (default to #FFFFFF10). Note that the actual image should have transparency.
+
 ## 2.5.0
 
 - Allow to choose underline horizontal alignment ([#46](https://github.com/Dalvany/dalvany-image-panel/issues/46)).
